Add number abbreviation to StyledTabs
